Create serializers module.
