+++ users/resources.py
@@ -1,38 +1,9 @@
-from rest_framework import serializers, status
-from rest_framework.decorators import api_view
+from rest_framework import status
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
-from users.models import Student, CustomUser, Prize
-
-
-class StudentSerializer(serializers.Serializer):
-
-    pk = serializers.IntegerField(read_only=True)
-    email = serializers.EmailField(source="user.email")
-    username = serializers.CharField(source="user.username")
-    first_name = serializers.CharField(source="user.first_name")
-    last_name = serializers.CharField(source="user.last_name")
-    total_points = serializers.IntegerField()
-
-    def create(self, validated_data):
-        user_data = validated_data.pop("user")
-        user = CustomUser.objects.create(**user_data)
-
-        return Student.objects.create(user=user, total_points=validated_data["total_points"])
-
-    def update(self, instance, validated_data):
-        user = instance.user
-        user_data = validated_data.get("user", {})
-        user.email = user_data.get("email", user.email)
-        user.username = user_data.get("username", user.username)
-        user.first_name = user_data.get("first_name", user.first_name)
-        user.last_name = user_data.get("last_name", user.last_name)
-        user.save()
-
-        instance.total_points = validated_data.get("total_points", instance.total_points)
-        instance.save()
-        return instance
+from users.models import Student, Prize
+from users.serializers import StudentSerializer, PrizeSerializer
 
 
 class StudentsResource(APIView):
@@ -70,13 +41,7 @@ class StudentsResource(APIView):
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class PrizeSerializer(serializers.Serializer):
-    pk = serializers.IntegerField()
-    name = serializers.CharField()
-
-
 class PrizesResource(APIView):
-
     def get(self, request, pk=None):
         if pk is None:
             prizes = Prize.objects.all()
